requirements.txt
amqp==5.1.1
asgiref==3.5.0
autopep8==2.0.1
beautifulsoup4==4.12.3
billiard==3.6.4.0
boto3==1.26.146
botocore==1.29.146
celery==5.2.7
certifi==2021.10.8
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.11
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cryptography==41.0.3
Deprecated==1.2.13
distlib==0.3.6
Django==4.0.2
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-cors-headers==3.7.0
django-redis==5.2.0
django-timezone-field==5.0
django-translations==1.3.0
djangorestframework==3.13.1
djangorestframework-api-key==2.1.0
docopt==0.6.2
filelock==3.12.0
flatbuffers==23.5.26
gunicorn==20.1.0
h11==0.9.0
hpack==3.0.0
hyperframe==5.2.0
idna==3.3
jmespath==1.0.1
kombu==5.2.4
libclang==16.0.0
ndg-httpsclient==0.5.1
numpy==1.24.3
packaging==21.3
pipenv==2023.6.12
platformdirs==3.5.1
prompt-toolkit==3.0.36
psycopg2==2.9.3
psycopg2-binary==2.8.5
pyasn1==0.4.8
pycodestyle==2.10.0
pycparser==2.21
PyJWT==1.7.1
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyparsing==3.0.7
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
redis==4.1.3
requests==2.27.1
rfc3986==1.5.0
s3transfer==0.6.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.2
stripe==2.73.0
tomli==2.0.1
types-cryptography==3.3.23.1
tzdata==2021.5
urllib3==1.26.8
vine==5.0.0
virtualenv==20.23.0
virtualenv-clone==0.5.7
wcwidth==0.2.6
wrapt==1.13.3
